Tentative fixes for SBCL compiler noise/some missed optimizations. (#1)
* Tentative fixes for SBCL compiler noise/some missed optimizations.

Before this PR, loading luckless with a clear cache and ASDF verbosity on,
results in a lot of output, ending with:

;;   caught 1 WARNING condition
;;   caught 6 STYLE-WARNING conditions
;;   printed 26 notes

After this PR is applied, the noise is much less, and ends with:

;;   caught 3 STYLE-WARNING conditions
;;   printed 7 notes

Most of the noise were STYLE-WARNINGs which I mostly ignored, for stuff like
&optional and &key in the same lambda list.

Some things were just missing type declarations.

Some things I was not sure about, and I noted them in comments. I temporarily
locally declared the speed optimization quality back to 1 in the few places it
was unhappy about types, generating generic code, that should be looked into.

In one particular case, a constant variable defined in hashtable.lisp was being
referenced in cat.lisp, even though the file ordering prevented there even being
a variable at compile time. I moved that one variable to the other file.

There is more work to do, and this definitely shouldn't be merged as is, but
it's a start.

Fixes.

It turns out I was of by a day on my monthly SBCL release, and tested again on
2.4.1. The results were similar, and all of my changes apply still.

However, I removed the locally optimize changes, which contributes to 3 extra
NOTEs being emitted.

How the author wants to solve the generic code there is up to them, and just
forcing a lower optimization is one option, but, in an effort to clean this PR
up with something reasonable in all cases, I left them for another issue.

That brings us to:

;   caught 3 STYLE-WARNING conditions
;   printed 10 notes

(I reverted the previous commit and squashed it locally, force-pushing to the
remote fork this PR is based off of)

* I'm a moron and forgot to unstage the locally hacks before the last commit.

The last commit with the locally hacks gives as compiler output:

;   caught 3 STYLE-WARNING conditions
;   printed 5 notes

This latest commit without the locally hacks, is:

;   caught 3 STYLE-WARNING conditions
;   printed 10 notes